refactor: moved strategy
