+++ src/lib/src/core/viewer-service/calculate-next-page-factory.ts
@@ -1,11 +1,10 @@
-import {
-  CalculateNextPageStrategy,
-  DashboardModeCalculateNextPageStrategy,
-  PageModeCalculateNextPageStrategy
-} from './calculate-next-page-strategy';
+import { CalculateNextPageStrategy } from './calculate-next-page-strategy';
 import { ViewerMode } from '../models/viewer-mode';
+import { DashboardModeCalculateNextPageStrategy } from './dashboard-mode-calculate-next-page-strategy';
+import { PageModeCalculateNextPageStrategy } from './page-mode-calculate-next-page-strategy';
 
 export class CalculateNextPageFactory {
+
   public static create(mode: ViewerMode): CalculateNextPageStrategy {
     if (mode === ViewerMode.DASHBOARD) {
       return new DashboardModeCalculateNextPageStrategy();
