+++ src/lib/src/core/viewer-service/page-mode-calculate-next-page-strategy.ts
@@ -1,0 +1,8 @@
+import { CalculateNextPageStrategy, NextPageCriteria } from './calculate-next-page-strategy';
+
+export class PageModeCalculateNextPageStrategy implements CalculateNextPageStrategy {
+
+  calculateNextPage(criteria: NextPageCriteria): number {
+    return 0;
+  }
+}
